Fix sign comparison in cl_cin

+++ src/client/cl_cin.c
@@ -125,7 +125,7 @@ typedef struct {
 
 	long samplesPerPixel;                               // defaults to 2
 	byte*               gray;
-	unsigned int xsize, ysize, maxsize, minsize;
+	long xsize, ysize, maxsize, minsize;
 
 	qboolean half, smootheddouble, inMemory;
 	long normalBuffer0;
